WIP Patch
1) Downloading data only once
2) Likert scale for ordinal values

+++ ctrl4ai/datasets.py
@@ -7,6 +7,8 @@ Created on Tue May 19 19:41:40 2020
 
 import pandas as pd
 import requests
+import os
+
 pd.set_option('mode.chained_assignment', None)
 
 def titanic():
@@ -17,8 +19,10 @@ def titanic():
   Description: Sample Dataset
   Returns: Pandas DataFrame
   """
-  titanic_request = requests.get("https://github.com/vkreat-tech/ctrl4ai/raw/master/ctrl4ai/sample_datasets/titanic.csv.gz", allow_redirects=True)
-  open('titanic.csv', 'wb').write(titanic_request.content)
+  if not os.path.exists('titanic.csv'):
+    print('Downloading titanic.csv ......')
+    titanic_request = requests.get("https://github.com/vkreat-tech/ctrl4ai/raw/master/ctrl4ai/sample_datasets/titanic.csv.gz", allow_redirects=True)
+    open('titanic.csv', 'wb').write(titanic_request.content)
   dataset=pd.read_csv('titanic.csv',compression='gzip')
   dataset=dataset.drop(['Unnamed: 0'],axis=1)
   return dataset
@@ -32,8 +36,10 @@ def trip_fare():
   Description: Sample Dataset
   Returns: Pandas DataFrame
   """
-  trip_fare_request = requests.get("https://github.com/vkreat-tech/ctrl4ai/raw/master/ctrl4ai/sample_datasets/trip_fare.csv.gz", allow_redirects=True)
-  open('trip_fare.csv', 'wb').write(trip_fare_request.content)
+  if not os.path.exists('trip_fare.csv'):
+    print('Downloading trip_fare.csv ......')
+    trip_fare_request = requests.get("https://github.com/vkreat-tech/ctrl4ai/raw/master/ctrl4ai/sample_datasets/trip_fare.csv.gz", allow_redirects=True)
+    open('trip_fare.csv', 'wb').write(trip_fare_request.content)
   dataset = pd.read_csv('trip_fare.csv',compression='gzip',parse_dates=['pickup_datetime','dropoff_datetime'])
   dataset=dataset.drop(['Unnamed: 0'],axis=1)
   return dataset
